v5.0: carry over manual edits from 3.9
A file with a large enumerated list.  Likely to be added to
in every new version of libclang.  So even though comments in this
file are very misleading and making the code hard to read, it's
probably better to keep it as close to the generated file as possible
for now.