v5.0: add test for TUResourceUsage that fails
The TUResourceUsage struct is one of those types that shares its c
buffer with a slice that is returned to the user, even though the user
is expected to call Dispose().  This can lead to a use after free
situation, which while not causing a crash in this test, does result in
the backing store of the slice having been changed under the covers.

The next commit will fix this by reworking the API for getting the slice
ResourceUsages.