v5.0: Add ExceptionSpecification() methods
Both Cursor and Type get the ExceptionSpecification() method.

This is a little more dangerous than most commits because the names
of the enum and clang calls didn't exactly line up.

The enum name started with Cursor_ and had a *Kind_* in the middle
but there were no functions using or returning these enums.

The two clang functions, one taking a Type, the other a Cursor,
had names that ended in *Type.  So do the libclang authors consider
the enum a Type or a Kind?

I tried to simplify things for the Go users by removing the Cursor_
from the beginning and the Kind from the middle of the enum name.

And I created an enum for the -1 value.  Again a little dangerous.
One of the clang functions specifically said it returns a -1 if
a non function type was passed in.  But the one taking a cursor just
implied the result would not be valid.  So will it always return -1
in that case?