v5.0: change Remapping struct and related API
The Remapping struct is now a pair of strings, Original and Transformed.

There are now just two Remapping related functions:
NewRemappings() and NewRemappingsFromFileList().

Each of the two functions returns a slice of Remappings.
All the memory returned to the user comes from the Go domain and can
be garbage collected appropriately.
The user no longer had a dispose to call for the Remapping struct
because no C memory is retained.