v5.0: bug fix: C.CXString cleanup in PlatformAvailability type
The C.CXPlatformAvailability struct has two members of
type C.CXString, and the dispose defined for this struct,
C.clang_disposeCXPlatformAvailability() calls the dispose
routines on its own members, rightly so.

So our getters for the Platform and Message strings should not
be calling dispose on those CXStrings.

Probably would lead to duplicate frees.