v5.0: cleanup Cursor.OverriddenCursors()
The pointer returned by the C routine is usually going to be nil
so skip the work of creating a slice and then copying to another slice.
Just allow a nil slice to be returned.  Idiomatic.

Also added warning that this even works because the Go Cursor type is a
struct that is the same size as the C.CXCursor struct.