Add cmd/go-clang-includes example
This shows being able to pass arguments to go-clang (actually libclang)
without having to tease out the filename, and how the include filenames
can be found.

The extra sugar printed by the other commands is not included with this
one as the ability to list include files is nice to have in general.