cleanup GlobalOptFlags
Fix the constants so they are all of the proper type.

Make the set and get functions use the proper Go type.

    Index.GlobalOptions()
    Index.SetGlobalOptions()

Remove Spelling() for this type, just use String().

Fix the String() function to allow for fact that the bits can be
bitwise ORed together and there could be unexpected bits.

Create unit tests for the set and get functions.