Switch to using stringer to create String() methods.
Call to stringer is in stringer.go.

It is one long line of arguments to go:generate stringer because
a single call to stringer takes 15 seconds given the number of
cgo files that need to be processed. By doing all in one shot,
that package load time is not incurred multiple times and the result
is stored in a a single file, named stringer_string.go.