Create back "pointers" in AST and improve test
A new constant cursorkind.Back is added and a new type
of cursor is added to the ast to enable tracking of
clang cursors that are visited multiple times. The clang
ast allows a child cursor to appear in multiple spots in the tree.

Update the test code to use a table from table_test.go which itself
can be generated by the test. So table_test.got can be moved to
table_test.go whenever new functionality or new tests warrant.