Add -update to go test
'go test' now does not create the .got file unless the .go version
is different. So an extra file read in each test, but no longer
a file write in each test. Only write the .got file when it is
different.

And if running 'go test -update', don't treat the diff as an error.

'go test -v -update' is useful to see the diff and have the file
updated and see the test pass for all other reasons.